+++ HISTORY.rst
@@ -1,7 +1,13 @@
 Release History
 ===============
 
-.. Unreleased Changes
+Unreleased Changes
+------------------
+* ``pygeoprocessing.zonal_statistics`` will now optionally include a count of
+  the number of pixels per value encountered under each polygon. A warning
+  will be logged when invoked on floating-point rasters, as using this on
+  continuous rasters can result in excessive memory consumption. To use this
+  feature, set ``include_value_counts=True`` when calling ``zonal_statistics``.
 
 2.3.4 (2022-08-22)
 ------------------